LayoutTests/editing/deleting/smart-delete-001.html

<!DOCTYPE html>
<html> 
<head>

<style>
.editing { 
    border: 2px solid red; 
    font-size: 24px; 
}
.explanation { 
    border: 2px solid blue; 
    padding: 12px; 
    font-size: 24px; 
    margin-bottom: 24px;
}
.scenario { margin-bottom: 16px;}
.scenario:first-line { font-weight: bold; margin-bottom: 16px;}
.expected-results:first-line { font-weight: bold }
</style>
<script src="../editing.js"></script>
<script src="../../resources/ui-helper.js"></script>

<script>

async function editingTest() {
    await UIHelper.selectWordByDoubleTapOrClick(document.getElementById('test'));
    deleteCommand();
}

</script>

<title>Editing Test</title> 
</head> 
<body>

<div class="explanation">
<div class="scenario">
Tests: 
<br>
Smart delete when deleting the first word on a line.
</div>
<div class="expected-results">
Expected Results:
<br>
The first word and the space following the first word should be deleted. It should like this this:
<BR>
bar baz
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="test" class="editing">
foo bar baz
</div>
</div>

<script>
runDumpAsTextEditingTest(true);
</script>

</body>
</html>
